Refactored structure of project
