+++ src/apdu_handlers.h
@@ -1,9 +1,12 @@
 #ifndef SKYCOIN_APDU_HANDLERS_H
 #define SKYCOIN_APDU_HANDLERS_H
 
-#include "os.h"
-#include "sky.h"
+#include "ui/newTransaction.h"
+#include "skycoin-api/skycoin_crypto.h"
 #include "ux.h"
+
+#include "os.h"
+#include "cx.h"
 
 /** Instructions start **/
 #define INS_RESET 0x00                  // instruction to reset
@@ -53,14 +56,15 @@ void handleGetPublicKey(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t da
 void handleGetVersion(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t dataLength, volatile unsigned int *flags,
                       volatile unsigned int *tx);
 
-void handleGetSignedPublicKey(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t dataLength, volatile unsigned int *flags,
+void
+handleGetSignedPublicKey(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t dataLength, volatile unsigned int *flags,
                          volatile unsigned int *tx);
 
 void handleGetAddress(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t dataLength, volatile unsigned int *flags,
                       volatile unsigned int *tx);
 
 void handleSignTxn(uint8_t p1, uint8_t p2, uint8_t *dataBuffer, uint16_t dataLength, volatile unsigned int *flags,
-                      volatile unsigned int *tx);
+                   volatile unsigned int *tx);
 
 #endif //SKYCOIN_APDU_HANDLERS_H
