+++ src/skycoin-api/skycoin_crypto.h
@@ -3,7 +3,7 @@
 
 #include "os.h"
 #include "cx.h"
-#include "ux.h"
+#include "../ux.h"
 #include <stdbool.h>
 #include "os_io_seproxyhal.h"
