--- src/txn.c
@@ -1,1 +1,0 @@
-
