--- src/util_funcs.h
@@ -1,11 +1,0 @@
-//
-// Created by midren on 3/12/19.
-//
-
-#ifndef LEDGER_NANO_UTIL_FUNCS_H
-#define LEDGER_NANO_UTIL_FUNCS_H
-
-// Return 1 if they are equal
-int mem_equal(const void *arr1, const void *arr2, unsigned int len);
-
-#endif //LEDGER_NANO_UTIL_FUNCS_H
